ptn de quotes
