+++ creat_list.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   creat_list.c                                       :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: david <david@student.42.fr>                +#+  +:+       +#+        */
+/*   By: dchellen <dchellen@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/02/22 12:14:26 by david             #+#    #+#             */
-/*   Updated: 2025/02/24 20:09:58 by david            ###   ########.fr       */
+/*   Updated: 2025/02/25 18:11:15 by dchellen         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
