+++ creat_token.c
@@ -6,7 +6,7 @@
 /*   By: dchellen <dchellen@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/02/10 16:02:21 by dchellen          #+#    #+#             */
-/*   Updated: 2025/02/25 16:03:23 by dchellen         ###   ########.fr       */
+/*   Updated: 2025/02/25 18:11:15 by dchellen         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -25,15 +25,17 @@ int	creat_tokken(char *input, t_shell *shell)
 		shell->creat.begin = shell->creat.i;
 
 		// gestion des doubles quotes
-		int result_d = double_quotes(input, &shell->creat.i);
-		int result_s = single_quotes(input, &shell->creat.i);
-		if (result_d == ERROR || result_s == ERROR)
+		shell->creat.result_d = double_quotes(input, &shell->creat.i);
+		shell->creat.result_s = single_quotes(input, &shell->creat.i);
+		if (shell->creat.result_d == ERROR || shell->creat.result_s == ERROR)
 			return (ERROR);
-		else if (result_d == VALID || result_s == VALID)
+		else if (shell->creat.result_d == VALID)
 			shell->creat.end = shell->creat.i;
+		else if (shell->creat.result_s == VALID)
+			shell->creat.end = shell->creat.i;			
 
 		// gestion des redirections
-		else if (detect_redirections(input, &shell->creat.i) == ERROR)
+		if (detect_redirections(input, &shell->creat.i) == ERROR)
 			return (ERROR);
 		else if (detect_redirections(input, &shell->creat.i) == VALID)
 		{
