+++ main.c
@@ -6,7 +6,7 @@
 /*   By: dchellen <dchellen@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/02/22 12:34:48 by david             #+#    #+#             */
-/*   Updated: 2025/02/25 12:35:12 by dchellen         ###   ########.fr       */
+/*   Updated: 2025/02/25 18:11:12 by dchellen         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
